Move _CopyOptionsBase to _OptionsBase
