+++ src/abaqus/_OptionsBase.py
@@ -2,6 +2,9 @@ import inspect
 
 
 class _OptionsBase:
+    """
+    This is a base class for all options classes.
+    """
 
     @classmethod
     def get_user_attributes(cls) -> list[str]:
@@ -26,3 +29,29 @@ class _OptionsBase:
         """
         for key, value in kwargs.items():
             setattr(self, key, value)
+
+
+class _CopyOptionsBase(_OptionsBase):
+    """
+    This is a base class for options classes where the first argument is the options object used to copy options
+    in the setValues method.
+    """
+
+    def setValues(self, options, **kwargs):
+        """This method modifies the CopyOptions object.
+
+        Parameters
+        ----------
+        options : CopyOptions
+            An object from which values are to be copied. If other arguments are also supplied to setValues,
+            they will override the values in **options**. The default value is None.
+        kwargs
+            Keyword arguments.
+        """
+        if options is not None:
+            for key in self.get_user_attributes():
+                setattr(self, key, getattr(options, key))
+
+        for key, value in kwargs:
+            if value is not None:
+                setattr(self, key, value)
